Added the grayscale and negate tests
On Challenge — Explore the “water.jpg” picture in the images folder. Write a method
fixUnderwater() to modify the pixel colors to make the fish easier to see. Create a class
(static) method to test this new method in the class PictureTester. Be sure to call the new
test method in the main method in PictureTester.